requirements.txt
pillow
pyheif

